requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
streamlit==1.32.0
requests==2.31.0
